+++ src/relay/backend/graph_runtime_codegen.cc
@@ -28,15 +28,12 @@
 #include <tvm/relay/expr_functor.h>
 #include <tvm/runtime/device_api.h>
 
-#include <cstdint>
-#include <cstdio>
 #include <list>
 #include <string>
 #include <vector>
 
 #include "compile_engine.h"
 #include "utils.h"
-#include "../../runtime/json/json_node.h"
 
 namespace tvm {
 namespace relay {
@@ -628,6 +625,12 @@ TVM_REGISTER_GLOBAL("relay.build_module._GraphRuntimeCodegen")
 
 namespace dmlc {
 namespace json {
+// JSON utils
+template <typename T>
+inline bool SameType(const dmlc::any& data) {
+  return std::type_index(data.type()) == std::type_index(typeid(T));
+}
+
 template <>
 struct Handler<std::shared_ptr<tvm::relay::backend::GraphNode>> {
   inline static void Write(dmlc::JSONWriter* writer,
@@ -636,6 +639,63 @@ struct Handler<std::shared_ptr<tvm::relay::backend::GraphNode>> {
   }
   inline static void Read(dmlc::JSONReader* reader,
                           std::shared_ptr<tvm::relay::backend::GraphNode>* data) {
+    LOG(FATAL) << "Not implemented.";
+  }
+};
+template <>
+struct Handler<std::unordered_map<std::string, dmlc::any>> {
+  inline static void Write(dmlc::JSONWriter* writer,
+                           const std::unordered_map<std::string, dmlc::any>& data) {
+    writer->BeginObject();
+    for (const auto& kv : data) {
+      auto k = kv.first;
+      const dmlc::any& v = kv.second;
+      if (SameType<std::string>(v)) {
+        writer->WriteObjectKeyValue(k, dmlc::get<std::string>(v));
+      } else if (SameType<int>(v)) {
+        writer->WriteObjectKeyValue(k, dmlc::get<int>(v));
+      } else if (SameType<std::vector<size_t>>(v)) {
+        writer->WriteObjectKeyValue(k, dmlc::get<std::vector<size_t>>(v));
+      } else if (SameType<std::vector<std::vector<int64_t>>>(v)) {
+        writer->WriteObjectKeyValue(k, dmlc::get<std::vector<std::vector<int64_t>>>(v));
+      } else if (SameType<std::vector<std::string>>(v)) {
+        writer->WriteObjectKeyValue(k, dmlc::get<std::vector<std::string>>(v));
+      } else if (SameType<std::vector<dmlc::any>>(v)) {
+        writer->WriteObjectKeyValue(k, dmlc::get<std::vector<dmlc::any>>(v));
+      } else {
+        LOG(FATAL) << "Not supported";
+      }
+    }
+    writer->EndObject();
+  }
+  inline static void Read(dmlc::JSONReader* reader,
+                          std::unordered_map<std::string, dmlc::any>* data) {
+    LOG(FATAL) << "Not implemented.";
+  }
+};
+
+template <>
+struct Handler<std::vector<dmlc::any>> {
+  inline static void Write(dmlc::JSONWriter* writer, const std::vector<dmlc::any>& data) {
+    writer->BeginArray();
+    for (const auto& v : data) {
+      if (SameType<std::string>(v)) {
+        writer->WriteArrayItem(dmlc::get<std::string>(v));
+      } else if (SameType<int>(v)) {
+        writer->WriteArrayItem(dmlc::get<int>(v));
+      } else if (SameType<std::vector<size_t>>(v)) {
+        writer->WriteArrayItem(dmlc::get<std::vector<size_t>>(v));
+      } else if (SameType<std::vector<std::vector<int64_t>>>(v)) {
+        writer->WriteArrayItem(dmlc::get<std::vector<std::vector<int64_t>>>(v));
+      } else if (SameType<std::vector<std::string>>(v)) {
+        writer->WriteArrayItem(dmlc::get<std::vector<std::string>>(v));
+      } else {
+        LOG(FATAL) << "Not supported";
+      }
+    }
+    writer->EndArray();
+  }
+  inline static void Read(dmlc::JSONReader* reader, std::vector<dmlc::any>* data) {
     LOG(FATAL) << "Not implemented.";
   }
 };
